requirements.txt
Flask==3.1.2
gunicorn==23.0.0

easyocr==1.7.2
torch==2.9.1
torchvision==0.24.1

numpy==2.2.6
pillow==12.0.0
opencv-python-headless==4.12.0.88
